Move the no-provider auto-skip into a dedicated effect component
